Updated typings up to 538a30bb921c01e777b0a49b4899af21b933d30e
